Move AudioSettings into audio_settings module
