+++ noita-proxy/src/app.rs
@@ -1,0 +1,1 @@
+
